Chore (Windows): move `perflib_.h` to somewhere global
